@@ -5,7 +5,10 @@ Math functions.
 from __future__ import annotations
 
 import math
-from typing import Sequence
+from typing import TYPE_CHECKING
+
+if TYPE_CHECKING:
+    from typing import Sequence
 
 
 def gcd(*numbers: int) -> int:
