@@ -46,7 +46,11 @@ def regrep(
     """
     compiled = {k: re.compile(v) for k, v in patterns.items()}
     matches = collections.defaultdict(list)
-    gen = reverse_readfile(filename) if reverse else zopen(filename, "rt")
+    gen = (
+        reverse_readfile(filename)
+        if reverse
+        else zopen(filename, "rt", encoding="utf-8")
+    )
     for i, line in enumerate(gen):
         for k, p in compiled.items():
             if m := p.search(line):
