@@ -9,6 +9,7 @@ from pathlib import Path
 import pytest
 
 from monty.io import (
+    EncodingWarning,
     FileLock,
     FileLockException,
     _get_line_ending,
@@ -371,25 +372,102 @@ class TestReverseReadfile:
 
 
 class TestZopen:
-    def test_zopen(self):
-        with zopen(os.path.join(TEST_DIR, "myfile_gz.gz"), mode="rt") as f:
-            assert f.read() == "HelloWorld.\n\n"
-        with zopen(os.path.join(TEST_DIR, "myfile_bz2.bz2"), mode="rt") as f:
-            assert f.read() == "HelloWorld.\n\n"
-        with zopen(os.path.join(TEST_DIR, "myfile_bz2.bz2"), "rt") as f:
-            assert f.read() == "HelloWorld.\n\n"
-        with zopen(os.path.join(TEST_DIR, "myfile_xz.xz"), "rt") as f:
-            assert f.read() == "HelloWorld.\n\n"
-        with zopen(os.path.join(TEST_DIR, "myfile_lzma.lzma"), "rt") as f:
-            assert f.read() == "HelloWorld.\n\n"
-        with zopen(os.path.join(TEST_DIR, "myfile"), mode="rt") as f:
-            assert f.read() == "HelloWorld.\n\n"
+    @pytest.mark.parametrize("extension", [".txt", ".bz2", ".gz", ".xz", ".lzma"])
+    def test_read_write_files(self, extension):
+        """Test read/write in binary/text mode:
+        - uncompressed text file: .txt
+        - compressed files: bz2/gz/xz/lzma
+        """
+        filename = f"test_file{extension}"
+        content = "This is a test file.\n"
 
-    def test_Path_objects(self):
-        p = Path(TEST_DIR) / "myfile_gz.gz"
+        with ScratchDir("."):
+            # Test write and read in text mode
+            with zopen(filename, "wt", encoding="utf-8") as f:
+                f.write(content)
 
-        with zopen(p, mode="rt") as f:
-            assert f.read() == "HelloWorld.\n\n"
+            with zopen(Path(filename), "rt", encoding="utf-8") as f:
+                assert f.read() == content
+
+            # Test write and read in binary mode
+            with zopen(Path(filename), "wb") as f:
+                f.write(content.encode())
+
+            with zopen(filename, "rb") as f:
+                assert f.read() == content.encode()
+
+    def test_lzw_files(self):
+        """gzip is not really able to (de)compress LZW files.
+
+        TODO: remove text file real_lzw_file.txt.Z after dropping
+        ".Z" extension support
+        """
+        # Test a fake LZW file (just with .Z extension but DEFLATED algorithm)
+        filename = "test.Z"
+        content = "This is not a real LZW compressed file.\n"
+
+        with (
+            ScratchDir("."),
+            pytest.warns(FutureWarning, match="compress LZW-compressed files"),
+        ):
+            # Test write and read in text mode
+            with zopen(filename, "wt", encoding="utf-8") as f:
+                f.write(content)
+
+            with zopen(filename, "rt", encoding="utf-8") as f:
+                assert f.read() == content
+
+            # Test write and read in binary mode
+            with zopen(filename, "wb") as f:
+                f.write(content.encode())
+
+            with zopen(filename, "rb") as f:
+                assert f.read() == content.encode()
+
+        # Cannot decompress a real LZW file
+        with (
+            pytest.raises(gzip.BadGzipFile, match="Not a gzipped file"),
+            zopen(f"{TEST_DIR}/real_lzw_file.txt.Z", "rt", encoding="utf-8") as f,
+        ):
+            f.read()
+
+    @pytest.mark.parametrize("extension", [".txt", ".bz2", ".gz", ".xz", ".lzma"])
+    def test_warnings(self, extension):
+        filename = f"test_warning{extension}"
+        content = "Test warning"
+
+        with ScratchDir("."):
+            # Default `encoding` warning
+            with (
+                pytest.warns(EncodingWarning, match="use UTF-8 by default"),
+                zopen(filename, "wt") as f,
+            ):
+                f.write(content)
+
+            # Implicit text/binary `mode` warning
+            warnings.filterwarnings(
+                "ignore", category=EncodingWarning, message="argument not specified"
+            )
+            with (
+                pytest.warns(
+                    FutureWarning, match="discourage using implicit binary/text"
+                ),
+                zopen(filename, "r") as f,
+            ):
+                if extension == ".txt":
+                    assert f.readline() == content
+                else:
+                    assert f.readline().decode("utf-8") == content
+
+            # Implicit `mode` warning
+            with (
+                pytest.warns(FutureWarning, match="discourage using a default `mode`"),
+                zopen(filename) as f,
+            ):
+                if extension == ".txt":
+                    assert f.readline() == content
+                else:
+                    assert f.readline().decode("utf-8") == content
 
 
 class TestFileLock:
