@@ -88,8 +88,8 @@ class TestMakedirs_p:
         makedirs_p(self.test_dir_path)
         assert os.path.exists(self.test_dir_path)
         makedirs_p(self.test_dir_path)
-        with pytest.raises(OSError):
-            makedirs_p(os.path.join(TEST_DIR, "myfile_txt"))
+        with pytest.raises(OSError, match="exists"):
+            makedirs_p(os.path.join(TEST_DIR, "3000_lines.txt"))
 
     def teardown_method(self):
         os.rmdir(self.test_dir_path)
